Abstracted a new `DirectionAware` component out
